requirements.txt
faiss-cpu
langchain-text-splitters
numpy
